refactor(projects): перенести фильтры в отдельный компонент
Jira issue: VEGA-599
