+++ src/pages/projects/ProjectsFilter/index.ts
@@ -1,0 +1,1 @@
+export { ProjectsFilter } from './ProjectsFilter';
