+++ src/pages/projects/ProjectsPageView.tsx
@@ -1,15 +1,6 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
-import {
-  Button,
-  ChoiceGroup,
-  IconBookmarkFilled,
-  IconProps,
-  IconSearch,
-  Loader,
-  Text,
-  TextField,
-} from '@gpn-prototypes/vega-ui';
+import { Button, Loader, Text } from '@gpn-prototypes/vega-ui';
 
 import { ProjectUpdateType } from '../../__generated__/types';
 
@@ -19,79 +10,8 @@ import { ProjectsTable } from './ProjectsTable';
 
 import './ProjectsPage.css';
 
-type Item = {
-  name: string;
-  icon?: React.FC<IconProps>;
-};
-
-const filterItems = [
-  {
-    name: 'Мои',
-  },
-  {
-    name: 'Избранные',
-    icon: IconBookmarkFilled,
-  },
-  {
-    name: 'Последние',
-  },
-  {
-    name: 'Все',
-  },
-] as Item[];
-
-type TextFieldOnChangeArgs = {
-  value: string | null;
-  name?: string;
-  e: React.ChangeEvent;
-  id?: string | number;
-};
-
-type ProjectFilterType = {
-  onInputSearch(value: string | null): void;
-  onChangeFilter(value: Item | null): void;
-};
-
 const testId = {
   root: 'ProjectsPage:root',
-};
-
-const ProjectFilter: React.FC<ProjectFilterType> = ({ onInputSearch, onChangeFilter }) => {
-  const [searchValue, setSearchValue] = React.useState<string | null>(null);
-  const [filterValue, setFilterValue] = React.useState<Item | null>(null);
-
-  const handleFilter = ({ value }: { value: Item | null }): void => {
-    setFilterValue(value);
-    onChangeFilter(value);
-  };
-
-  return (
-    <div className={cn('Filter')}>
-      <ChoiceGroup<Item>
-        size="s"
-        value={filterValue}
-        items={filterItems}
-        getLabel={(item): string => item.name}
-        getIcon={(item): React.FC<IconProps> | undefined => item.icon}
-        onChange={handleFilter}
-        name="ChoiceGroup"
-        multiple={false}
-      />
-      <div className={cn('FilterField')}>
-        <TextField
-          value={searchValue}
-          onChange={({ value }: TextFieldOnChangeArgs): void => {
-            setSearchValue(value);
-            onInputSearch(value);
-          }}
-          size="s"
-          leftSide={IconSearch}
-          width="full"
-          placeholder="Введите название проекта или имя автора"
-        />
-      </div>
-    </div>
-  );
 };
 
 type Props = {
@@ -137,14 +57,14 @@ export const ProjectsPageView: React.FC<Props> = (props) => {
             <Button label="Создать новый проект" size="s" />
           </Link>
         </div>
-        <div className={cn('Projects')}>
-          <ProjectFilter
-            // eslint-disable-next-line no-console
-            onInputSearch={(value: string): void => console.log(`InputSearch: ${value}`)}
-            // eslint-disable-next-line no-console
-            onChangeFilter={(value: Item): void => console.log(`Filter: ${value.name}`)}
-          />
-        </div>
+
+        {/* <ProjectsFilter
+          // eslint-disable-next-line no-console
+          onInputSearch={(value: string): void => console.log(`InputSearch: ${value}`)}
+          // eslint-disable-next-line no-console
+          onChangeFilter={(value: Item): void => console.log(`Filter: ${value.name}`)}
+        /> */}
+
         {props.isLoading ? (
           <div className={cn('Loader')}>
             <Loader />
